# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriSmart API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface CreateConversationBody {
  title: string;
}

export interface SendMessageBody {
  content: string;
}

export interface AnalyzeCropDiseaseBody {
  imageBase64: string;
  /** @nullable */
  cropType?: string | null;
}

export interface DiseaseAnalysisResult {
  diseaseName: string;
  confidence: number;
  severity: string;
  description: string;
  symptoms: string[];
  treatment: string[];
  prevention: string[];
  affectedCrop: string;
}

export interface ForecastDay {
  date: string;
  tempMin: number;
  tempMax: number;
  condition: string;
  precipitation: number;
}

export interface WeatherAdvisory {
  location: string;
  temperature: number;
  humidity: number;
  windSpeed: number;
  condition: string;
  description: string;
  advisories: string[];
  farmingTips: string[];
  forecast: ForecastDay[];
}

export interface MarketPrice {
  id: number;
  cropName: string;
  variety: string;
  state: string;
  market: string;
  minPrice: number;
  maxPrice: number;
  modalPrice: number;
  unit: string;
  date: string;
  trend: string;
  changePercent: number;
}

export interface MarketPriceSummary {
  totalCrops: number;
  avgPriceChange: number;
  topGainers: MarketPrice[];
  topLosers: MarketPrice[];
  lastUpdated: string;
}

export interface GovernmentScheme {
  id: number;
  name: string;
  category: string;
  description: string;
  benefits: string;
  eligibility: string;
  applicationProcess: string;
  /** @nullable */
  deadline?: string | null;
  isActive: boolean;
  ministry: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  helpline?: string | null;
}

export interface DashboardSummary {
  totalSchemes: number;
  activeSchemes: number;
  cropCount: number;
  recentDetections: number;
  weatherAlert: boolean;
  marketHighlight: string;
}

export type GetWeatherAdvisoryParams = {
  city?: string;
  lat?: number;
  lon?: number;
};

export type ListMarketPricesParams = {
  crop?: string;
  state?: string;
};

export type ListGovernmentSchemesParams = {
  category?: string;
  search?: string;
};