# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgriSmart API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /government-schemes:
    get:
      operationId: listGovernmentSchemes
      tags: [government-schemes]
      summary: List government agricultural schemes
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of schemes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GovernmentScheme"
  /government-schemes/{id}:
    get:
      operationId: getGovernmentScheme
      tags: [government-schemes]
      summary: Get a single government scheme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scheme details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernmentScheme"
        "404":
          description: Not found
...
components:
  schemas:
...
    GovernmentScheme:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        benefits:
          type: string
        eligibility:
          type: string
        applicationProcess:
          type: string
        deadline:
          type: ["string", "null"]
        isActive:
          type: boolean
        ministry:
          type: string
        website:
          type: ["string", "null"]
        helpline:
          type: ["string", "null"]
      required:
        - id
        - name
        - category
        - description
        - benefits
        - eligibility
        - applicationProcess
        - isActive
        - ministry
...